Change type of progress_bar_update_rate to int32_t

+++ rosbag2_transport/include/rosbag2_transport/play_options.hpp
@@ -138,12 +138,12 @@ public:
   // read_ahead_queue_size value to buffer (and order) more messages.
   MessageOrder message_order = MessageOrder::RECEIVED_TIMESTAMP;
 
-  // Rate in Hz at which to print progress bar.
-  double progress_bar_update_rate = 3.0;
+  // Progress bar update rate in times per second (Hz)
+  int32_t progress_bar_update_rate = 3;
 
   // Number of separation lines to print in between the playback output
   // and the progress bar.
-  int progress_bar_separation_lines = 3;
+  int32_t progress_bar_separation_lines = 3;
 };
 
 }  // namespace rosbag2_transport
